# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ETask API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/runner/delete/{id}": {
      "delete": {
        "tags": [
          "执行单元"
        ],
        "summary": "删除执行单元",
        "description": "权限标识: delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/detail/{id}": {
      "get": {
        "tags": [
          "执行单元"
        ],
        "summary": "执行单元详情",
        "description": "权限标识: get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/list": {
      "post": {
        "tags": [
          "执行单元"
        ],
        "summary": "执行单元列表",
        "description": "权限标识: view",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRunnerReq"
              },
              "example": {
                "keyword": "search_key",
                "kind": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/list/by_ids": {
      "post": {
        "tags": [
          "执行单元"
        ],
        "summary": "批量查询执行单元",
        "description": "权限标识: ids",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRunnerByIDsReq"
              },
              "example": {
                "ids": [
                  1,
                  2,
                  3
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/list/exclude_codebook_id": {
      "post": {
        "tags": [
          "执行单元"
        ],
        "summary": "复用执行单元",
        "description": "权限标识: view_exclude_codebook_id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListExcludeCodebookIDReq"
              },
              "example": {
                "codebook_id": 1,
                "keyword": "search_key",
                "kind": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/list/{codebook_id}": {
      "get": {
        "tags": [
          "执行单元"
        ],
        "summary": "当前绑定执行单元",
        "description": "权限标识: view_runners",
        "parameters": [
          {
            "name": "codebook_id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/register": {
      "post": {
        "tags": [
          "执行单元"
        ],
        "summary": "注册执行单元",
        "description": "权限标识: add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRunnerReq"
              },
              "example": {
                "codebook_id": 1,
                "codebook_secret": "admin",
                "desc": "相关业务描述信息",
                "handler": "string",
                "kind": "string",
                "name": "示例名称",
                "parameter_defaults": "string",
                "program_kind": "string",
                "tags": [
                  "item1",
                  "item2"
                ],
                "target": "string",
                "variables": [
                  "item1",
                  "item2"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/runner/update": {
      "post": {
        "tags": [
          "执行单元"
        ],
        "summary": "更新执行单元",
        "description": "权限标识: edit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunnerReq"
              },
              "example": {
                "codebook_id": 1,
                "codebook_secret": "admin",
                "desc": "相关业务描述信息",
                "handler": "string",
                "id": 1,
                "kind": "string",
                "name": "示例名称",
                "parameter_defaults": "string",
                "program_kind": "string",
                "tags": [
                  "item1",
                  "item2"
                ],
                "target": "string",
                "variables": [
                  "item1",
                  "item2"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ListExcludeCodebookIDReq": {
        "type": "object",
        "properties": {
          "codebook_id": {
            "type": "integer",
            "example": 1
          },
          "keyword": {
            "type": "string",
            "example": "search_key"
          },
          "kind": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "codebook_id": 1,
          "keyword": "search_key",
          "kind": "string"
        }
      },
...
      "ListRunnerByIDsReq": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              1,
              2,
              3
            ]
          }
        },
        "example": {
          "ids": [
            1,
            2,
            3
          ]
        }
      },
      "ListRunnerReq": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "example": "search_key"
          },
          "kind": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "keyword": "search_key",
          "kind": "string"
        }
      },
...
      "RegisterRunnerReq": {
        "type": "object",
        "properties": {
          "codebook_id": {
            "type": "integer",
            "example": 1
          },
          "codebook_secret": {
            "type": "string",
            "example": "admin"
          },
          "desc": {
            "type": "string",
            "example": "相关业务描述信息"
          },
          "handler": {
            "type": "string",
            "example": "string"
          },
          "kind": {
            "type": "string",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          },
          "parameter_defaults": {
            "type": "string",
            "example": "string"
          },
          "program_kind": {
            "type": "string",
            "example": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          },
          "target": {
            "type": "string",
            "example": "string"
          },
          "variables": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          }
        },
        "example": {
          "codebook_id": 1,
          "codebook_secret": "admin",
          "desc": "相关业务描述信息",
          "handler": "string",
          "kind": "string",
          "name": "示例名称",
          "parameter_defaults": "string",
          "program_kind": "string",
          "tags": [
            "item1",
            "item2"
          ],
          "target": "string",
          "variables": [
            "item1",
            "item2"
          ]
        }
      },
...
      "UpdateRunnerReq": {
        "type": "object",
        "properties": {
          "codebook_id": {
            "type": "integer",
            "example": 1
          },
          "codebook_secret": {
            "type": "string",
            "example": "admin"
          },
          "desc": {
            "type": "string",
            "example": "相关业务描述信息"
          },
          "handler": {
            "type": "string",
            "example": "string"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "kind": {
            "type": "string",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          },
          "parameter_defaults": {
            "type": "string",
            "example": "string"
          },
          "program_kind": {
            "type": "string",
            "example": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          },
          "target": {
            "type": "string",
            "example": "string"
          },
          "variables": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          }
        },
        "example": {
          "codebook_id": 1,
          "codebook_secret": "admin",
          "desc": "相关业务描述信息",
          "handler": "string",
          "id": 1,
          "kind": "string",
          "name": "示例名称",
          "parameter_defaults": "string",
          "program_kind": "string",
          "tags": [
            "item1",
            "item2"
          ],
          "target": "string",
          "variables": [
            "item1",
            "item2"
          ]
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}