# internal/web/runner/handler.go
package runner

import (
	"errors"

	"github.com/Duke1616/eiam/pkg/web/capability"
	"github.com/Duke1616/etask/internal/domain"
	"github.com/Duke1616/etask/internal/errs"
	"github.com/Duke1616/etask/internal/pkg/security"
	runnerSvc "github.com/Duke1616/etask/internal/service/runner"
	"github.com/Duke1616/etask/pkg/contract/model"
	"github.com/Duke1616/etask/pkg/contract/perm"
	"github.com/Duke1616/etask/pkg/cryptox"
	"github.com/ecodeclub/ekit/slice"
	"github.com/ecodeclub/ginx"
	"github.com/gin-gonic/gin"
)

var _ ginx.Handler = &Handler{}

type Handler struct {
	svc runnerSvc.Service
	capability.IRegistry
}

func NewHandler(svc runnerSvc.Service) *Handler {
	return &Handler{
		svc:       svc,
		IRegistry: capability.NewRegistry("task", "runner", "执行单元"),
	}
}

func (h *Handler) PublicRoutes(_ *gin.Engine) {
}

func (h *Handler) IdentifyRoutes(_ *gin.Engine) {
}

func (h *Handler) PrivateRoutes(server *gin.Engine) {
	g := server.Group("/api/runner")
	g.POST("/register", h.Define("注册执行单元", "add").
		Bind(ginx.B[RegisterRunnerReq](h.Register)),
	)
	g.POST("/list", h.Define("执行单元列表", "view").
		Needs(perm.Runner.Ids, perm.Codebook.ViewRunners, perm.Runner.ViewExcludeCodebookId).
		Bind(ginx.B[ListRunnerReq](h.List)),
	)
	g.GET("/detail/:id", h.Define("执行单元详情", "get").
		Bind(ginx.W(h.Detail)),
	)
	g.POST("/update", h.Define("更新执行单元", "edit").
		Needs(perm.Runner.Get).
		Bind(ginx.B[UpdateRunnerReq](h.Update)),
	)
	g.DELETE("/delete/:id", h.Define("删除执行单元", "delete").
		Bind(ginx.W(h.Delete)),
	)
	g.POST("/list/by_ids", h.Define("批量查询执行单元", "ids").
		NoSync().
		Bind(ginx.B[ListRunnerByIDsReq](h.ListByIDs)),
	)
	g.GET("/list/:codebook_id", h.For(model.Codebook).Define("当前绑定执行单元", "view_runners").
		NoSync().
		Bind(ginx.W(h.ListByCodebookID)),
	)
	g.POST("/list/exclude_codebook_id", h.Define("复用执行单元", "view_exclude_codebook_id").
		NoSync().
		Bind(ginx.B[ListExcludeCodebookIDReq](h.ListExcludeCodebookID)),
	)
}

func (h *Handler) Register(ctx *ginx.Context, req RegisterRunnerReq) (ginx.Result, error) {
	id, err := h.svc.Create(ctx, h.toDomain(req))
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: id}, nil
}

func (h *Handler) List(ctx *ginx.Context, req ListRunnerReq) (ginx.Result, error) {
	rs, total, err := h.svc.List(ctx, req.Offset, req.Limit, req.Keyword, req.Kind)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: h.toListResp(rs, total)}, nil
}

func (h *Handler) Detail(ctx *ginx.Context) (ginx.Result, error) {
	id, err := ctx.Param("id").AsInt64()
	if err != nil {
		return invalidIDResult, err
	}
	r, err := h.svc.FindByID(ctx, id)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: h.toVO(r)}, nil
}

func (h *Handler) Update(ctx *ginx.Context, req UpdateRunnerReq) (ginx.Result, error) {
	r, err := h.toUpdateDomain(ctx, req)
	if err != nil {
		return h.translateError(err), err
	}
	count, err := h.svc.Update(ctx, r)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: count}, nil
}

func (h *Handler) Delete(ctx *ginx.Context) (ginx.Result, error) {
	id, err := ctx.Param("id").AsInt64()
	if err != nil {
		return invalidIDResult, err
	}
	count, err := h.svc.Delete(ctx, id)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: count}, nil
}

func (h *Handler) ListByIDs(ctx *ginx.Context, req ListRunnerByIDsReq) (ginx.Result, error) {
	rs, err := h.svc.ListByIDs(ctx, req.IDs)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: h.toListResp(rs, int64(len(rs)))}, nil
}

func (h *Handler) ListByCodebookID(ctx *ginx.Context) (ginx.Result, error) {
	codebookID, err := ctx.Param("codebook_id").AsInt64()
	if err != nil {
		return invalidIDResult, err
	}
	rs, err := h.svc.ListByCodebookID(ctx, codebookID)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: h.toExecutionListResp(rs)}, nil
}

func (h *Handler) ListExcludeCodebookID(ctx *ginx.Context, req ListExcludeCodebookIDReq) (ginx.Result, error) {
	rs, total, err := h.svc.ListExcludeCodebookID(ctx, req.Offset, req.Limit, req.CodebookID, req.Keyword, req.Kind)
	if err != nil {
		return h.translateError(err), err
	}
	return ginx.Result{Msg: "success", Data: h.toListResp(rs, total)}, nil
}

func (h *Handler) translateError(err error) ginx.Result {
	if errors.Is(err, errs.ErrInvalidParameter) {
		return invalidParameterResult(err)
	}
	return systemErrorResult
}

func (h *Handler) toDomain(req RegisterRunnerReq) domain.Runner {
	return domain.Runner{
		Name:              req.Name,
		CodebookSecret:    req.CodebookSecret,
		CodebookID:        req.CodebookID,
		ProgramKind:       domain.ProgramKind(req.ProgramKind),
		Tags:              req.Tags,
		Kind:              domain.RunnerKind(req.Kind),
		Variables:         h.toVariablesDomain(req.Variables),
		ParameterDefaults: req.ParameterDefaults,
		Action:            domain.RunnerActionRegistered,
		Target:            req.Target,
		Handler:           req.Handler,
		Desc:              req.Desc,
	}
}

func (h *Handler) toUpdateDomain(ctx *ginx.Context, req UpdateRunnerReq) (domain.Runner, error) {
	old, err := h.svc.FindByID(ctx, req.ID)
	if err != nil {
		return domain.Runner{}, err
	}
	oldVars := slice.ToMap(old.Variables, func(element domain.RunnerVariable) string {
		return element.Key
	})
	return domain.Runner{
		ID:                req.ID,
		Name:              req.Name,
		CodebookSecret:    req.CodebookSecret,
		CodebookID:        req.CodebookID,
		ProgramKind:       domain.ProgramKind(req.ProgramKind),
		Tags:              req.Tags,
		Kind:              domain.RunnerKind(req.Kind),
		Variables:         h.toUpdateVariablesDomain(oldVars, req.Variables),
		ParameterDefaults: req.ParameterDefaults,
		Action:            domain.RunnerActionRegistered,
		Target:            req.Target,
		Handler:           req.Handler,
		Desc:              req.Desc,
	}, nil
}

func (h *Handler) toVariablesDomain(req []Variable) []domain.RunnerVariable {
	return slice.Map(req, func(_ int, src Variable) domain.RunnerVariable {
		return domain.RunnerVariable{
			Key:    src.Key,
			Secret: src.Secret,
			Value:  src.Value,
		}
	})
}

func (h *Handler) toUpdateVariablesDomain(oldVars map[string]domain.RunnerVariable, req []Variable) []domain.RunnerVariable {
	return slice.Map(req, func(_ int, src Variable) domain.RunnerVariable {
		value := src.Value
		if src.Secret {
			val, ok := oldVars[src.Key]
			if ok && (src.Value == "" || src.Value == cryptox.DefaultMask) {
				value = val.Value
			}
		}
		return domain.RunnerVariable{
			Key:    src.Key,
			Secret: src.Secret,
			Value:  value,
		}
	})
}

func (h *Handler) toListResp(rs []domain.Runner, total int64) ListRunnersResp {
	return ListRunnersResp{
		Total: total,
		Runners: slice.Map(rs, func(_ int, src domain.Runner) RunnerVO {
			return h.toVO(src)
		}),
	}
}

func (h *Handler) toExecutionListResp(specs []domain.RunnerExecutionSpec) ListRunnersResp {
	return ListRunnersResp{
		Total: int64(len(specs)),
		Runners: slice.Map(specs, func(_ int, spec domain.RunnerExecutionSpec) RunnerVO {
			result := h.toVO(spec.Runner)
			result.Variables = h.toVariablesVO(spec.Variables)
			return result
		}),
	}
}

func (h *Handler) toVO(req domain.Runner) RunnerVO {
	return RunnerVO{
		ID:                req.ID,
		Name:              req.Name,
		Kind:              req.Kind.String(),
		CodebookID:        req.CodebookID,
		ProgramKind:       string(req.ProgramKind),
		Tags:              req.Tags,
		Desc:              req.Desc,
		Target:            req.Target,
		Handler:           req.Handler,
		CTime:             req.CTime,
		UTime:             req.UTime,
		Variables:         h.toVariablesVO(req.Variables),
		ParameterDefaults: req.ParameterDefaults,
	}
}

func (h *Handler) toVariablesVO(variables []domain.RunnerVariable) []Variable {
	masked := security.NewVariableMasker().MaskVariables(variables)
	return slice.Map(masked, func(_ int, src domain.RunnerVariable) Variable {
		return Variable{Key: src.Key, Secret: src.Secret, Value: src.Value}
	})
}